+++ internal/provider/resource_gitlab_tag_protection_test.go
@@ -409,7 +409,7 @@ func testAccCheckGitlabTagProtectionAttributes(pt *gitlab.ProtectedTag, want *te
 			remainingWantedAccessLevelsAllowedToCreate[int(api.AccessLevelNameToValue[v])] = struct{}{}
 		}
 
-		remainingWantedUserIDsAllowedToCreate := map[int]struct{}{}
+		remainingWantedUserIDsAllowedToCreate := map[int64]struct{}{}
 		for _, v := range want.UsersAllowedToCreate {
 			users, _, err := testutil.TestGitlabClient.Users.ListUsers(&gitlab.ListUsersOptions{
 				Username: gitlab.Ptr(v),
@@ -422,7 +422,7 @@ func testAccCheckGitlabTagProtectionAttributes(pt *gitlab.ProtectedTag, want *te
 			}
 			remainingWantedUserIDsAllowedToCreate[users[0].ID] = struct{}{}
 		}
-		remainingWantedGroupIDsAllowedToCreate := map[int]struct{}{}
+		remainingWantedGroupIDsAllowedToCreate := map[int64]struct{}{}
 		for _, v := range want.GroupsAllowedToCreate {
 			group, _, err := testutil.TestGitlabClient.Groups.GetGroup(v, nil)
 			if err != nil {
